Fix addChannelToDefaultCategory
