+++ app/actions/local/category.ts
@@ -3,9 +3,15 @@
 
 import {Model} from '@nozbe/watermelondb';
 
+import {CHANNELS_CATEGORY, DMS_CATEGORY} from '@constants/categories';
 import DatabaseManager from '@database/manager';
 import {prepareCategories, prepareCategoryChannels, queryCategoriesByTeamIds, getCategoryById} from '@queries/servers/categories';
+import {getCurrentUserId} from '@queries/servers/system';
+import {queryMyTeams} from '@queries/servers/team';
+import {isDMorGM} from '@utils/channel';
 import {pluckUnique} from '@utils/helpers';
+
+import type ChannelModel from '@typings/database/models/servers/channel';
 
 export const deleteCategory = async (serverUrl: string, categoryId: string) => {
     const database = DatabaseManager.serverDatabases[serverUrl]?.database;
@@ -109,3 +115,49 @@ export const toggleCollapseCategory = async (serverUrl: string, categoryId: stri
         return {error};
     }
 };
+
+export async function addChannelToDefaultCategory(serverUrl: string, channel: Channel | ChannelModel, prepareRecordsOnly = false) {
+    const operator = DatabaseManager.serverDatabases[serverUrl]?.operator;
+    if (!operator) {
+        return {error: `${serverUrl} database not found`};
+    }
+
+    const {database} = operator;
+
+    const teamId = 'teamId' in channel ? channel.teamId : channel.team_id;
+    const userId = await getCurrentUserId(database);
+    if (!userId) {
+        return {error: 'no current user id'};
+    }
+
+    const models: Model[] = [];
+    const categoriesWithChannels: CategoryWithChannels[] = [];
+
+    if (isDMorGM(channel)) {
+        const allTeamIds = await queryMyTeams(database).fetchIds();
+        const categories = await queryCategoriesByTeamIds(database, allTeamIds).fetch();
+        const channelCategories = categories.filter((c) => c.type === DMS_CATEGORY);
+        for await (const cc of channelCategories) {
+            const cwc = await cc.toCategoryWithChannels();
+            cwc.channel_ids.unshift(channel.id);
+            categoriesWithChannels.push(cwc);
+        }
+    } else {
+        const categories = await queryCategoriesByTeamIds(database, [teamId]).fetch();
+        const channelCategory = categories.find((c) => c.type === CHANNELS_CATEGORY);
+        if (channelCategory) {
+            const cwc = await channelCategory.toCategoryWithChannels();
+            cwc.channel_ids.unshift(channel.id);
+            categoriesWithChannels.push(cwc);
+        }
+
+        const ccModels = await prepareCategoryChannels(operator, categoriesWithChannels);
+        models.push(...ccModels);
+    }
+
+    if (models.length && !prepareRecordsOnly) {
+        await operator.batchRecords(models);
+    }
+
+    return {models};
+}
