+++ app/utils/categories.ts
@@ -1,10 +1,6 @@
 // Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
 // See LICENSE.txt for license information.
 
-export function makeCategoryId(name: string, userId: string, teamId: string) {
-    return `${name}_${userId}_${teamId}`;
-}
-
 export function makeCategoryChannelId(teamId: string, channelId: string) {
     return `${teamId}_${channelId}`;
 }
